Add a credential plugin that uses GitHub Apps to get tokens
This patch uses PyGitHub to implement it. It currently only supports
GitHub App IDs due to limitations in PyGitHub [[1]]. It also
takes care of passing it as a string into the library because PyGitHub
does not perform type conversion, which sometimes leads to producing
invalid JWTs [[2]] as the underlying library PyJWT lacks validation
[[3]].

The plugin accepts both GitHub App and installation IDs as integers
and normalizes them into strings.

The tests use ephemerally-produced in-memory RSA keys for JWT
verification. As an optimization, they are 1024-bit sized because it
is cheaper and makes the tests more performant compared to bigger
values.

As a temporary measure, the patch includes a suppression of WPS202
in flake8 for `github_app.py`.

A few type-ignores were added around PyGitHub exception handling due
to them using `Any` [[4]].

The patch also includes a change to the flake8-eradicate config
letting distinguish comments with URL lists.

[1]: https://github.com/PyGithub/PyGithub/issues/3213
[2]: https://github.com/PyGithub/PyGithub/issues/3214
[3]: https://github.com/jpadilla/pyjwt/issues/1039
[4]: https://github.com/PyGithub/PyGithub/issues/3218

Co-authored-by: Sviatoslav Sydorenko <webknjaz@redhat.com>