🚑 Suppress insufficient coverage pytest warning
This is emitted by `pytest-cov` [[1]] and is turned into an error by
the default `-Werror` passed to Python, not the `filterwarnings`
setting within the `pytest`.

The patch selectively suppresses the warning so there's nothing to
turn into an error in the first place. Insufficient coverage still
marks the test session as failed as it's supposed to.

[1]: https://github.com/pytest-dev/pytest-cov/issues/675