🧪 Add Ruff linter to pre-commit
It is integrated with auto-fixing enabled which effectively replaces
`pyupgrade`, `isort` and `docformatter`. It does not yet fully [[1]]
cover `autopep8`, though.

`autopep8` is being removed since the combination of
`add-trailing-comma` and `ruff` works rather well with trailing
commas.

And the formatter mode isn't enabled either due to partial
incompatibility with `add-trailing-comma`.

The WPS flake8 plugin is upgraded to v1.1.0 as it drops a lot of
checks performed by `ruff`. Other plugins are not removed and should be
evaluated for being covered by `ruff` in follow-ups.

[1]: https://github.com/astral-sh/ruff/issues/5564#issuecomment-1623901448